+++ assets/shared/templates/calendar/utils/calendar.scss
