+++ assets/shared/templates/calendar/utils/helper.js
