+++ assets/shared/templates/rss/utils/rss.scss
