+++ assets/shared/templates/slideshow/utils/slideshow.scss
