+++ assets/shared/templates/table/utils/table.scss
