+++ assets/shared/templates/travel/utils/travel.scss
@@ -1,6 +1,6 @@
 @font-face {
   font-family: "rubics";
-  src: url("./fonts/rubik-regular.woff") format("woff");
+  src: url("fontsubik-regular.woff") format("woff");
 }
 
 .grid {
